requirements.txt
aiohttp
async-timeout
attrs
chardet
idna
multidict
typing-extensions
websockets
yarl
PyJWT
motor